Allow running all phpt tests without configuration
This commit adds a --EXTENSIONS-- section to the gd and imagick
tests. This way those tests will require no additional configuration
when running 'make test' on a system where the gd and/or imagick
extensions are installed.

On old versions of PHP5 where run-tests.php does not support the
extensions section, the section is ignored.